+++ CHANGELOG.md
@@ -26,6 +26,7 @@ For information on changes in released versions of Teku, see the [releases page]
  - Improved discv5 compliance
  - Changed the builder `is online\is offline` logs to `is available\is not available`
  - Updated REST API to use Javalin 5
+ - Added `/eth/v1/beacon/states/{state_id}/randao` to beacon-api.
 
 ### Bug Fixes
  - Fix missing status filters (active, pending, exited, withdrawal) for the `/eth/v1/beacon/states/{state_id}/validators` endpoint